refactor(web): extract book expand modal logic into custom hook
